server/forecast/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
prophet
python-dateutil
pymongo 
python-dotenv
scikit-learn